refactored snap/ directory into static classes w/associated tests
