+++ packages/hedera-wallet-snap/packages/snap/src/rpc/account/deleteAccount.ts
@@ -20,9 +20,9 @@
 
 import { providerErrors } from '@metamask/rpc-errors';
 import { heading, text } from '@metamask/snaps-ui';
-import { createHederaClient } from '../../snap/account';
+import { HederaClientFactory } from '../../snap/HederaClientFactory';
 import { SnapUtils } from '../../utils/SnapUtils';
-import { updateSnapState } from '../../snap/state';
+import { SnapState } from '../../snap/SnapState';
 import { Account, AccountInfo } from '../../types/account';
 import { AccountBalance, TxReceipt } from '../../types/hedera';
 import { DeleteAccountRequestParams } from '../../types/params';
@@ -75,7 +75,7 @@ export async function deleteAccount(
       throw providerErrors.userRejectedRequest();
     }
 
-    const hederaClient = await createHederaClient(
+    const hederaClient = await HederaClientFactory.create(
       curve,
       privateKey,
       hederaAccountId,
@@ -97,7 +97,7 @@ export async function deleteAccount(
     // eslint-disable-next-line require-atomic-updates
     state.accountState[hederaEvmAddress][network].accountInfo =
       {} as AccountInfo;
-    await updateSnapState(state);
+    await SnapState.updateState(state);
   } catch (error: any) {
     const errMessage = `Error while trying to delete account: ${String(error)}`;
     console.error(errMessage);
