+++ packages/hedera-wallet-snap/packages/snap/src/rpc/transactions/transferCrypto.ts
@@ -22,7 +22,7 @@ import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
 import { HederaServiceImpl } from '../../services/impl/hedera';
-import { createHederaClient } from '../../snap/account';
+import { HederaClientFactory } from '../../snap/HederaClientFactory';
 import { SnapUtils } from '../../utils/SnapUtils';
 import { AccountInfo } from '../../types/account';
 import { SimpleTransfer, TxReceipt } from '../../types/hedera';
@@ -242,7 +242,7 @@ export async function transferCrypto(
       throw providerErrors.userRejectedRequest();
     }
 
-    const hederaClient = await createHederaClient(
+    const hederaClient = await HederaClientFactory.create(
       state.accountState[hederaEvmAddress][network].keyStore.curve,
       state.accountState[hederaEvmAddress][network].keyStore.privateKey,
       hederaAccountId,
