--- packages/hedera-wallet-snap/packages/snap/src/snap/dapp.ts
@@ -1,48 +1,0 @@
-/*-
- *
- * Hedera Wallet Snap
- *
- * Copyright (C) 2023 Tuum Tech
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-import { WalletSnapState } from '../types/state';
-import { updateSnapState } from './state';
-
-/**
- * Function that lets you add a friendly dApp.
- *
- * @param state - WalletSnapState.
- * @param dapp - Dapp.
- */
-export async function addFriendlyDapp(state: WalletSnapState, dapp: string) {
-  state.snapConfig.dApp.friendlyDapps.push(dapp);
-  await updateSnapState(state);
-}
-
-/**
- * Function that removes a friendly dApp.
- *
- * @param state - WalletSnapState.
- * @param dapp - Dapp.
- */
-export async function removeFriendlyDapp(state: WalletSnapState, dapp: string) {
-  // FIXME: TEST IF YOU CAN REFERENCE FRIENDLY DAPS
-  // let friendlyDapps = state.snapConfig.dApp.friendlyDapps;
-  // friendlyDapps = friendlyDapps.filter((app) => app !== dapp);
-  state.snapConfig.dApp.friendlyDapps =
-    state.snapConfig.dApp.friendlyDapps.filter((app) => app !== dapp);
-  await updateSnapState(state);
-}
