--- packages/hedera-wallet-snap/packages/snap/src/snap/dialog.ts
@@ -1,32 +1,0 @@
-/*-
- *
- * Hedera Wallet Snap
- *
- * Copyright (C) 2023 Tuum Tech
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-import { WalletSnapState } from '../types/state';
-import { updateSnapState } from './state';
-
-/**
- * Function that toggles the disablePopups flag in the config.
- *
- * @param state - WalletSnapState.
- */
-export async function updatePopups(state: WalletSnapState) {
-  state.snapConfig.dApp.disablePopups = !state.snapConfig.dApp.disablePopups;
-  await updateSnapState(state);
-}
